working on class domain
