+++ apps/server/src/modules/class/domain/class.do.ts
@@ -119,24 +119,3 @@ export class Class extends DomainObject<ClassProps> {
 		return classFullName;
 	}
 }
-
-export class ClassFactory {
-	public static create(props: Partial<ClassProps>): Class {
-		const baseProps = this.getBaseProps();
-		const classProps = { ...baseProps, ...props };
-
-		return new Class(classProps);
-	}
-
-	private static getBaseProps(): ClassProps {
-		return {
-			id: '',
-			name: '',
-			schoolId: '',
-			createdAt: new Date(),
-			updatedAt: new Date(),
-			userIds: [],
-			teacherIds: [],
-		};
-	}
-}
