+++ apps/server/src/modules/class/domain/index.ts
@@ -1,1 +1,2 @@
 export * from './class.do';
+export * from './class.factory';
